Support not storing the input edgelist...
Sequentially on my laptop, re-generating the edge list imposes a 2x cost
but eliminates a large amount of memory use (which is exactly the change
in number of make_edge() calls).  This still needs worked into the spec
and the other implementations.  Not storing the edge list makes more
memory available for the reported results.